rename file StoreUsers.py -> store_users.py
